rafactor: 필터 선택 안되는 이슈 해결
